% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{loadDataFromGit}
\alias{loadDataFromGit}
\title{Load Data from a GitHub Raw URL}
\usage{
loadDataFromGit(githubURL)
}
\arguments{
\item{githubURL}{The full URL to the raw RDS file on GitHub.
This URL should start with "https://raw.githubusercontent.com/".
The correct URL can be obtained by navigating to the file on GitHub,
clicking "Raw", and then copying the URL from the browser's download manager.}
}
\value{
Returns the data loaded from the RDS file.
}
\description{
This function downloads and loads data stored in an RDS file hosted on GitHub,
specifically from a raw content URL. It is designed to ensure the URL is correct,
downloads the file, loads it into R, and then cleans up by deleting the downloaded file.
}
\examples{
\dontrun{
  # Example URL (this is a placeholder, replace with a real URL for actual use)
  url <-
  "https://raw.githubusercontent.com/kreutz-lab/
  dia-benchmarking/main/data/diaWorkflowResults_allDilutions.rds"
  data <- loadDataFromGit(url)
}

}
