% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{msb.applyFunctionWithSeed}
\alias{msb.applyFunctionWithSeed}
\title{msb.applyFunctionWithSeed}
\usage{
msb.applyFunctionWithSeed(functionName, seed = 123, ...)
}
\arguments{
\item{functionName}{Function name as string}

\item{seed}{Seed defined outside this function, default = 123}

\item{...}{further parameters passed to function}
}
\value{
Result of the function the seed is applied to (res) and seed (seed)
}
\description{
Function to apply seed to a function which has a
random component, and
reset to previous seed afterwards
}
\examples{
mtx <- matrix(runif(100000),nrow=1000)
randomCols <- msb.applyFunctionWithSeed(sample, seed = 83434,
x = 1:ncol(mtx), size= min(50, ncol(mtx)))


}
\keyword{seed}
