% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coefficient_operations.R
\name{msb.omicsGetLmCoefs}
\alias{msb.omicsGetLmCoefs}
\title{Extract Linear Model Coefficients for Omics Data}
\usage{
msb.omicsGetLmCoefs(
  mtx = NULL,
  DE_idx = NULL,
  groupDesign = rep(c(1, 2), each = (ncol(mtx)/2)),
  chunksize = min(nrow(mtx), 400)
)
}
\arguments{
\item{mtx}{A matrix of omics data where rows represent features and
columns represent samples.}

\item{DE_idx}{A vector indicating the indices of differentially
expressed features.}

\item{groupDesign}{A vector indicating group assignments
for each column in the matrix.}

\item{chunksize}{The number of rows to process in each chunk.
Defaults to the minimum of 400 or the number of rows in the matrix.}
}
\value{
A list containing the coefficients of the linear models.
}
\description{
This function extracts linear model coefficients from a provided matrix
representing omics data, such as proteomics or genomics,
by fitting linear models
in chunks.
}
\examples{
nFeatures <- 100
nCols <- 10
data <- stats::rnorm(nFeatures * nCols)
mtx <- matrix(data, nrow = nFeatures, ncol = nCols)
DE_idx <- sample(1:nFeatures, 0.2*nFeatures)
msb.omicsGetLmCoefs(mtx, DE_idx)
}
