% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{msb.predictFromLmCoefs}
\alias{msb.predictFromLmCoefs}
\title{Simulate Matrix from Linear Model Coefficients}
\usage{
msb.predictFromLmCoefs(
  groupDesign,
  DE_idx = NULL,
  nFeatures = 500,
  percentDE = 0.2,
  coefs,
  int.mean = 19,
  addNoise = T,
  rowScale = 1,
  rowSdScale = 1,
  colScale = 1,
  FCScale = 1
)
}
\arguments{
\item{groupDesign}{A vector indicating the group design for each sample.}

\item{DE_idx}{Optional vector of indices for differentially expressed features.}

\item{nFeatures}{Integer specifying the number of features (rows) in the simulated matrix.}

\item{percentDE}{Proportion of features that are differentially expressed.}

\item{coefs}{List of coefficients including \code{featureCoefs}, \code{sampleCoefs}, and \code{FCCoefs}.}

\item{int.mean}{Mean intensity value to be added to each simulated data point.}

\item{addNoise}{Logical indicating whether to add simulated noise to the matrix.}

\item{rowScale}{Scaling factor for row coefficients.}

\item{rowSdScale}{Scaling factor for the standard deviation of rows.}

\item{colScale}{Scaling factor for column coefficients.}

\item{FCScale}{Scaling factor for fold change coefficients.}
}
\value{
A matrix representing simulated intensity values.
}
\description{
This function simulates an intensity matrix based on specified linear model coefficients
for a given experimental design. It optionally adds noise to the simulated values.
}
\examples{
  groupDesign <- rep(1:2, each = 5)
  DE_idx <- sample(1:500, 100)
  coefs <- list(featureCoefs = stats::rnorm(500),
  sampleCoefs = stats::rnorm(10), FCCoefs = stats::rnorm(100))
  sim_matrix <- msb.predictFromLmCoefs(groupDesign, DE_idx = DE_idx, coefs = coefs)
}
