% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateData.R
\name{msb.simulateDataFromBenchmark}
\alias{msb.simulateDataFromBenchmark}
\title{Simulate Data from Benchmark Data or Coefficients}
\usage{
msb.simulateDataFromBenchmark(
  mtx = NULL,
  groupDesign_mtx = NULL,
  DE_idx = NULL,
  jointCoefs = NULL,
  int.mean = 19,
  nFeatures = 2000,
  nSamples = 10,
  nDE = NULL,
  npat = 10,
  groupDesign_new = rep(c(1, 2), each = nSamples/2),
  drawReplace = T
)
}
\arguments{
\item{mtx}{Optional matrix used to derive LM and DIMAR coefficients if not provided.}

\item{groupDesign_mtx}{Group design for the input matrix \code{mtx} if it exists.}

\item{DE_idx}{Optional indices for differentially expressed (DE) features.}

\item{jointCoefs}{list of coefficients from linear model and DIMAR. Either
estimated from input mtx or found in data/.}

\item{int.mean}{The mean intensity value added to simulated data.}

\item{nFeatures}{Number of features to simulate in the output matrix.}

\item{nSamples}{Number of samples to simulate in the output matrix.}

\item{nDE}{Number of DE features, computed as a proportion of \code{nFeatures} by default.}

\item{npat}{Number of patterns for missing value simulations, defaults to 10.}

\item{groupDesign_new}{Group design for the new simulated matrix. By default
first half of samples are assigned to group 1 and the second half to group 2.}

\item{drawReplace}{Logical indicating whether sampling of coefficients is with replacement.}
}
\value{
An array of simulated data with dimensions nFeatures x nSamples x npat,
with attributes detailing the used coefficients.
}
\description{
This function simulates proteomics data either based on a given count matrix
or by previously estimated coefficients.
It utilizes a linear model to estimate coefficients for the intensity values
and a logistic regression as implemented in DIMAR to estimate coefficients
for the amount of missing values in each feature.
The function is designed to estimate coefficients from benchmark data where
it is known which features are differentially expressed. If no benchmark data
is available, the function can still be used to simulate data based on coefficients
previously estimated on data presented in
Fröhlich, K., Brombacher, E., Fahrner, M. et al.
Benchmarking of analysis strategies for data-independent acquisition
proteomics using a large-scale dataset
comprising inter-patient heterogeneity.
Nat Commun 13, 2622 (2022). https://doi.org/10.1038/s41467-022-30094-0
}
\details{
If \code{lmCoefs} or \code{dimarCoefs} are not provided, and an input matrix \code{mtx} is available,
the function will compute these coefficients based on \code{mtx} and \code{DE_idx}.
Make sure to install and load the DIMAR package as it is crucial for the function's operation.
}
\examples{

\dontrun{
url <- "https://raw.githubusercontent.com/kreutz-lab/dia-benchmarking/main/data/diaWorkflowResults_allDilutions.rds"
data <-
loadDataFromGit(url)
exp.df <- subsetDIAWorkflowData(data = data, DIAWorkflow = "DIANN_DIANN_AI",
experimentalComparisonGroups = c("1-12","1-25"),
rowSubset = seq(1,1000), colSubset = c(seq(1,5),seq(24,28)))

exp.df <- exp.df[rowSums(exp.df, na.rm = TRUE) > 0, ]
DE_idx <- grep("ECOLI", rownames(exp.df))
  sim_data <- msb.simulateDataFromBenchmark(
   mtx = as.matrix(exp.df),
   DE_idx = DE_idx,
   nFeatures = 800,
   nSamples = 20,
   npat = 2)
}


}
