% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{subsetDIAWorkflowData}
\alias{subsetDIAWorkflowData}
\title{This function subsets proteomics data based on specified Data-Independent Acquisition (DIA) workflow and
experimental comparison groups. It also allows further subsetting by specific rows and columns.}
\usage{
subsetDIAWorkflowData(
  data,
  DIAWorkflow,
  experimentalComparisonGroups,
  rowSubset = NULL,
  colSubset = NULL
)
}
\arguments{
\item{data}{List of datasets keyed by DIA workflow names. Can be loaded with
loadDatarFromGit()}

\item{DIAWorkflow}{Character string specifying the DIA workflow to subset from \code{data}.}

\item{experimentalComparisonGroups}{Character vector of groups to include based on their
tags in column names which might include specific concentrations of proteins or control setups.
This is highly adapted to the data from https://github.com/kreutz-lab/dia-benchmarking/tree/main}

\item{rowSubset}{Optional numeric vector indicating the rows to be retained.}

\item{colSubset}{Optional numeric vector indicating the columns to be retained.}
}
\value{
A subset of the data frame corresponding to the specified DIA workflow and experimental groups.
Further subsets by rows and columns if specified.
}
\description{
This function subsets proteomics data based on specified Data-Independent Acquisition (DIA) workflow and
experimental comparison groups. It also allows further subsetting by specific rows and columns.
}
\details{
The function requires a valid \code{diaWorkflow} which should be one of the keys in \code{data}.
The \code{experimentalComparisonGroups} should match parts of the column names in the data frame
corresponding to the groups of interest. This function stops and throws an error
if the required workflow or comparison groups are not found or if multiple workflows are specified.
}
\examples{
\dontrun{
rownames <- paste0("feature_",1:20)
colnames <- c(sample(1:5,5), paste0("group1_",sample(6:10,5)), paste0("group2_",sample(6:10,5)))
  data_list <- list(
    DIANN_DIANN_AI = data.frame(matrix(ncol = 15, nrow = 20, dimnames = list(rownames, colnames))),
    DIANN_MaxQuant = data.frame(matrix(ncol = 15, nrow = 20, dimnames = list(rownames, colnames)))
  )
  DIAWorkflow <- "DIANN_DIANN_AI"
  groups <- c("control", "group2")
  subsetted_data <- subsetDIAWorkflowData(data = data_list, DIAWorkflow, groups)
}

}
